+++ EscapeRoom/EscapeRoom/src/Intro.cpp
@@ -58,7 +58,8 @@ void Intro::handleInput(sf::Event& event, sf::RenderWindow& window) {
         exit();
     }
     else if (event.type == sf::Event::MouseButtonPressed && event.mouseButton.button == sf::Mouse::Left) {
-        sf::Vector2f clickPosition(event.mouseButton.x, event.mouseButton.y);
+        sf::Vector2f clickPosition(static_cast<float>(event.mouseButton.x), static_cast<float>(event.mouseButton.y));
+
 
         if (skipButton.isClicked(clickPosition)) {
             AudioManager::getInstance().playSoundEffect("Click.ogg");
